Fix build and pre refactor plugin part
