+++ build.gradle
@@ -47,7 +47,7 @@ dependencies {
     minecraft "com.mojang:minecraft:${project.minecraft_version}"
     mappings "net.fabricmc:yarn:${project.yarn_mappings}:v2"
     modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
-    implementation fileTree(dir: '/media/crystalneko/BA106C158B1B58EF/javatest/mods/ctLibFabric/build/libs', includes: ['ctLibMod-0.0.15.jar'])
+    implementation fileTree("libs")
     compileOnly 'org.spigotmc:spigot-api:1.20.1-R0.1-SNAPSHOT'
     compileOnly 'net.luckperms:api:5.4'
     modImplementation "me.lucko:fabric-permissions-api:0.2-SNAPSHOT"
